icc fixes
+ Fixed unused warnings
+ Fixed format errors
+ Removed unused variables
+ ...

Tim


git-svn-id: https://lechuck.informatik.rwth-aachen.de:2342/svn/fawkes/trunk@539 fc82b16a-671c-0410-8732-9903a8d2dd0d

+++ src/firevision/fvutils/rectification/rectfile.h
@@ -46,13 +46,13 @@ class RectificationInfoFile
    */
   typedef std::vector<RectificationInfoBlock *> RectInfoBlockVector;
 
-  uint8_t       version();
+  unsigned int  version();
   uint64_t      guid();
   const char *  model();
   bool          is_big_endian();
   bool          is_little_endian();
 
-  uint8_t       num_blocks();
+  unsigned int  num_blocks();
 
   void add_rectinfo_block(RectificationInfoBlock *block);
   void write(const char *file_name);
